+++ HISTORY.rst
@@ -7,6 +7,8 @@ Pending Release
 ---------------
 
 * New release notes here
+* Now supports expressions with multiple weekdays and ordinal days - thanks
+  @jbondia.
 
 1.0.1 (2016-02-09)
 ------------------